src/kafka_consumer_producer/requirements.txt
joblib==1.1.0
mysql-connector-python==8.0.31
kafka-python==2.0.2
pandas==1.4.1
numpy==1.21.2
jsonschema==4.4.0
